Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main(){
-    int a,b,c;
+    float a,b,c;
     printf("");
-    scanf("%d",&a);
+    scanf("%f",&a);
     printf("");
-    scanf("%d",&b);
+    scanf("%f",&b);
     printf("");
-    scanf("%d",&c);
+    scanf("%f",&c);
     int avg = (a+b+c)/3;
-    printf("Average: %d\n", avg);
+    printf("Average: %.2f\n", avg);
     return 0;
 }
